requirements.txt
fastapi
uvicorn[standard]
python-dotenv
bcrypt
PyJWT
pydantic
supabase
openai
fpdf
pytest
httpx
pydantic-settings
